dont try to remove nonexisting dirs

+++ UpdatePC.sh
@@ -8,9 +8,15 @@ wstool init src permo_car_pc.rosinstall
 wstool update -t src
 
 # clean up nodes not needed in remote repo
+if [ -d "src/PadPub/wiimote" ]; then
 rm -r src/PadPub/ps3joy
+fi
+if [ -d "src/PadPub/wiimote" ]; then
 rm -r src/PadPub/spacenav_node
+fi
+if [ -d "src/PadPub/wiimote" ]; then
 rm -r src/PadPub/wiimote
+fi
 
 catkin build
 rm permo_car_pc.rosinstall
